Test fixes and refactoring
- Mock file writing for package metadata config
- Move GrpcMetadataGen task into package_metadata_tasks module
- Temporarily hard-code a config value that's incompatible between
  GrpcMetadataGen and Gapic metadata config
