+++ pipeline/pipelines/grpc_generation_pipeline.py
@@ -15,7 +15,8 @@
 """Pipelines that run gRPC codegen"""
 
 from pipeline.pipelines import code_generation_pipeline as code_gen
-from pipeline.tasks import protoc_tasks, publish_tasks, python_grpc_tasks
+from pipeline.tasks import (package_metadata_tasks, protoc_tasks,
+                            publish_tasks, python_grpc_tasks)
 from pipeline.utils import task_utils
 
 
@@ -64,7 +65,7 @@ class _PythonGrpcTaskFactory(GrpcTaskFactoryBase):
         return [python_grpc_tasks.PythonChangePackageTask,
                 protoc_tasks.ProtoDescGenTask,
                 protoc_tasks.ProtoAndGrpcCodeGenTask,
-                protoc_tasks.GrpcPackageMetadataGenTask]
+                package_metadata_tasks.GrpcPackageMetadataGenTask]
 
 
 class _GoGrpcTaskFactory(GrpcTaskFactoryBase):
