+++ test/test_pipeline_baseline.py
@@ -87,8 +87,10 @@ def _test_error(pipeline_name, language, config, pipeline_kwargs,
 
 
 @mock.patch('pipeline.utils.task_utils.get_gradle_task_output')
-@mock.patch('pipeline.tasks.protoc_tasks.'
+@mock.patch('pipeline.tasks.package_metadata_tasks.'
             'GrpcPackageMetadataGenTask._googleapis_dir')
+@mock.patch('pipeline.tasks.package_metadata_tasks.'
+            'PackageMetadataConfigGenTask._write_yaml')
 @mock.patch('pipeline.tasks.python_grpc_tasks.PythonChangePackageTask.execute')
 @mock.patch('subprocess.call')
 @mock.patch('subprocess.check_call')
@@ -97,8 +99,8 @@ def _test_error(pipeline_name, language, config, pipeline_kwargs,
 @mock.patch('os.chdir')
 def _test_baseline(pipeline_name, config, extra_args, baseline,
                    mock_chdir, mock_time, mock_check_output, mock_check_call,
-                   mock_call, mock_python_pkg_task, mock__googleapis_dir,
-                   mock_gradle_task):
+                   mock_call, mock_python_pkg_task, mock__write_yaml,
+                   mock__googleapis_dir, mock_gradle_task):
     reporoot = os.path.abspath('.')
 
     # Execute pipeline args
